Restructure page for additional parameters
Add additional fields for parameters that are accepted by the Synthea
service to alter the details of the generated patients.

Ensure the patient and the parameters are found in summary/details
pairings to hide them away when not in use.

Redo the Summary/Details pairings to prevent the patient display
from toggling on every click.